# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_Edu0liver_prototype-healthy-api_internal_modules_agent_dto.AgentResponse:
    properties:
      handover_enabled:
        type: boolean
      handover_keywords:
        items:
          type: string
        type: array
      id:
        type: string
      max_output_tokens:
        type: integer
      model:
        type: string
      name:
        type: string
      status:
        type: string
      system_prompt:
        type: string
      temperature:
        type: number
    type: object
  github_com_Edu0liver_prototype-healthy-api_internal_modules_agent_dto.CreateAgentRequest:
    properties:
      handover_enabled:
        type: boolean
      handover_keywords:
        items:
          type: string
        type: array
      max_output_tokens:
        type: integer
      model:
        type: string
      name:
        minLength: 2
        type: string
      status:
        enum:
        - active
        - draft
        type: string
      system_prompt:
        type: string
      temperature:
        maximum: 2
        minimum: 0
        type: number
    required:
    - name
    - system_prompt
    type: object
...
  github_com_Edu0liver_prototype-healthy-api_internal_modules_automation_dto.AutomationResponse:
    properties:
      agent_id:
        type: string
      channel_id:
        type: string
      debounce_seconds:
        type: integer
      fallback_message:
        type: string
      id:
        type: string
      is_active:
        type: boolean
    type: object
  github_com_Edu0liver_prototype-healthy-api_internal_modules_automation_dto.CreateAutomationRequest:
    properties:
      agent_id:
        type: string
      business_hours:
        additionalProperties: {}
        type: object
      channel_id:
        type: string
      debounce_seconds:
        maximum: 60
        minimum: 0
        type: integer
      fallback_message:
        type: string
      is_active:
        type: boolean
    required:
    - agent_id
    - channel_id
    type: object
  github_com_Edu0liver_prototype-healthy-api_internal_modules_automation_dto.UpdateAutomationRequest:
    properties:
      agent_id:
        type: string
      business_hours:
        additionalProperties: {}
        type: object
      debounce_seconds:
        maximum: 60
        minimum: 0
        type: integer
      fallback_message:
        type: string
      is_active:
        type: boolean
    type: object
  github_com_Edu0liver_prototype-healthy-api_internal_modules_channel_dto.ChannelResponse:
    properties:
      active_agent_id:
        type: string
      external_account_id:
        type: string
      id:
        type: string
      instance_name:
        type: string
      name:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
...
  github_com_Edu0liver_prototype-healthy-api_internal_modules_channel_dto.CreateChannelRequest:
    properties:
      name:
        type: string
      number:
        description: external account / phone (E.164)
        type: string
      type:
        enum:
        - whatsapp
        - instagram
        type: string
    required:
    - type
    type: object
...
  github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: SaaS multi-tenant white-label de atendimento automatizado e omnichannel
    (WhatsApp/Instagram) com IA + RAG.
  title: API
  version: "1.0"
paths:
  /agents:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_internal_modules_agent_dto.AgentResponse'
              type: array
            type: object
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "403":
          description: insufficient role
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "429":
          description: rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List agents
      tags:
      - agents
    post:
      consumes:
      - application/json
      parameters:
      - description: Agent
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_internal_modules_agent_dto.CreateAgentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_internal_modules_agent_dto.AgentResponse'
        "400":
          description: invalid request
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "403":
          description: insufficient role
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "429":
          description: rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create agent
      tags:
      - agents
...
  /automations:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_internal_modules_automation_dto.AutomationResponse'
              type: array
            type: object
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "403":
          description: insufficient role
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "429":
          description: rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List automations
      tags:
      - automations
    post:
      consumes:
      - application/json
      parameters:
      - description: Automation
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_internal_modules_automation_dto.CreateAutomationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_internal_modules_automation_dto.AutomationResponse'
        "400":
          description: invalid body, or channel/agent not found in tenant
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "403":
          description: insufficient role
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "409":
          description: an active automation already exists for this channel
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "429":
          description: rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create automation
      tags:
      - automations
  /automations/{id}:
    delete:
      parameters:
      - description: Automation ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: invalid id
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "403":
          description: insufficient role
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "404":
          description: automation not found
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "429":
          description: rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete automation
      tags:
      - automations
    get:
      parameters:
      - description: Automation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: invalid id
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "403":
          description: insufficient role
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "404":
          description: automation not found
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "429":
          description: rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get automation
      tags:
      - automations
    put:
      consumes:
      - application/json
      parameters:
      - description: Automation ID
        in: path
        name: id
        required: true
        type: string
      - description: Automation
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_internal_modules_automation_dto.UpdateAutomationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_internal_modules_automation_dto.AutomationResponse'
        "400":
          description: invalid id or body
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "403":
          description: insufficient role
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "404":
          description: automation not found
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "409":
          description: an active automation already exists for this channel
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "429":
          description: rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update automation
      tags:
      - automations
...
  /channels:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_internal_modules_channel_dto.ChannelResponse'
              type: array
            type: object
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "403":
          description: insufficient role
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "429":
          description: rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List channels
      tags:
      - channels
    post:
      consumes:
      - application/json
      parameters:
      - description: Channel
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_internal_modules_channel_dto.CreateChannelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_internal_modules_channel_dto.ChannelResponse'
        "400":
          description: unsupported channel type or invalid body
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "403":
          description: insufficient role
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "429":
          description: rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create channel
      tags:
      - channels
...
  /channels/{id}/connection-state:
    get:
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid id
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "401":
          description: missing or invalid token
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "403":
          description: insufficient role
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "404":
          description: channel not found
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "429":
          description: rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/github_com_Edu0liver_prototype-healthy-api_pkg_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get connection state
      tags:
      - channels
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"